Fix online table error + more
Fixed the online table to actually use the latest UID from the main table.
Renamed the update_db() func to update_online() to fit it's description.
Removed a little bit of redundant code and optimized SQL queries to save time.
Also added tasklist to help track what needs to be added and whatnot.