update database format
added a new 'utext' column to the motds  table to store the unformatted plaintext version of motds. This will allow for better searches for servers with lots of formatting, e.g a server with a color for each letter. Also added the ROW=COMPRESSED function and first insert of the NO_ICON string for the icons table as to allow for less space taken by icons and guarantee that UID 1 of the icons table is NO_ICON.